Move clearTimeoutsBetweenTests to its own file too
