+++ apps/test/unit-tests.js
@@ -5,7 +5,7 @@ import {
   throwOnConsoleErrorsEverywhere,
   throwOnConsoleWarningsEverywhere
 } from './util/throwOnConsole';
-import {clearTimeoutsBetweenTests} from './util/testUtils';
+import {clearTimeoutsBetweenTests} from './util/clearTimeoutsBetweenTests';
 
 var __karmaWebpackManifest__ = [];
